move warnings and TDP parsing outside of CPU class
