+++ codecarbon/emissions_tracker.py
@@ -81,10 +81,19 @@ class BaseEmissionsTracker(ABC):
         elif cpu.is_rapl_available():
             logger.info("CODECARBON : Tracking Intel CPU via RAPL interface")
             self._hardware.append(CPU.from_utils(self._output_dir, "intel_rapl"))
-
         else:
+            logger.warning(
+                "CODECARBON : No CPU tracking mode found. Falling back on CPU constant mode."
+            )
             logger.info("CODECARBON : Tracking using constant")
-            self._hardware.append(CPU.from_utils(self._output_dir, "constant"))
+            tdp = cpu.TDP().tdp
+            if tdp:
+                self._hardware.append(CPU.from_utils(self._output_dir, "constant", tdp))
+            else:
+                logger.warning(
+                    f"CODECARBON : Failed to match CPU TDP constant. Falling back on a global constant."
+                )
+                self._hardware.append(CPU.from_utils(self._output_dir, "constant"))
 
         # Run `self._measure_power` every `measure_power_secs` seconds in a background thread
         self._scheduler.add_job(
